% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GClat.R
\name{GClat}
\alias{GClat}
\title{Fit ICAR Spatial Gamma-Count Model}
\usage{
GClat(Y, ID, W = NULL, shapefile = NULL, covariate = NULL)
}
\arguments{
\item{Y}{Vector of response variables (counts).}

\item{ID}{Vector of indexes of regions (spatial units).}

\item{W}{Optional adjacency matrix representing spatial connections between regions.
If not provided, it can be generated from a shapefile using the `shapefile` argument.}

\item{shapefile}{Optional shapefile representing the study region. If provided, the adjacency matrix (`W`)
will be calculated from the shapefile.}

\item{covariate}{Optional matrix of covariates. If not provided, the function assumes the model
is intercept-only.}
}
\value{
An object of class "inla" representing the fitted ICAR spatial GC model.
        The object contains model estimates, diagnostics, and other results.
}
\description{
This function fits an ICAR spatial Gamma-Count (GC) model to a given dataset using the INLA package.
It constructs the formula based on the provided covariate data and ID variables, and fits the model using
the specified adjacency matrix (`W`) or a shapefile of the study region.
}
\examples{
# Generate data from the GC spatial regression model with lattice spatial dependency
W <- rAdj(500) # Generate a random adjacency matrix
DDl <- rGClat(n = 500, alpha = 1, beta0 = 0.3, beta = c(-0.5, 0.5), W = W, V = 1)

# Prepare the data
Y <- DDl$y
covariate <- DDl$covariate
ID <- DDl$ID

# Fit the spatial GC model
ResultGC <- GClat(Y = Y, ID = ID, covariate = covariate, W = W)

# Summary of the model fit
summary(ResultGC)

}
