% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rGCgeo.R
\name{rGCgeo}
\alias{rGCgeo}
\title{Generate Data from GC Spatial Regression Model with Geospatial Dependency}
\usage{
rGCgeo(
  n = n,
  alpha,
  beta0,
  beta,
  V = NULL,
  rho = 1,
  sigma = NULL,
  range = NULL
)
}
\arguments{
\item{n}{Integer. The number of knots (or spatial units) for which the data should be generated.}

\item{alpha}{Numeric. The dispersion parameter of the Gamma-Count model.}

\item{beta0}{Numeric. The intercept term for the model.}

\item{beta}{Numeric vector. The regression coefficients (fixed effects) for the model.}

\item{V}{Optional numeric. The variance of the spatial random effects for lattice data.}

\item{rho}{Optional numeric. The spatial correlation coefficient for the CAR model. Default is 1.}

\item{sigma}{Optional numeric. The variance of the spatial random effects for geospatial data with Matern covariance.}

\item{range}{Optional numeric. The range parameter for geospatial data with Matern covariance.}
}
\value{
A list containing the following components:
\describe{
  \item{covariate}{A matrix of covariates with the specified number of knots (`n`) and columns based on the length of `beta`.}
  \item{phi}{A vector of spatial random effects generated from the Matern covariance model.}
  \item{eta}{A vector representing the linear predictor, calculated as the dot product of the covariates and coefficients plus the spatial effects (`phi`).}
  \item{mu}{A vector of mean response values calculated as the product of `alpha` and the exponential of `eta`.}
  \item{y}{A vector of simulated count data based on the GC model and the mean response values (`mu`).}
  \item{ID}{A vector of knot IDs from 1 to `n`.}
}
}
\description{
This function generates spatially dependent count data based on the Gamma-Count (GC) spatial regression model.
It uses a specified geospatial dependency model with parameters such as `sigma` for variance and `range` for spatial range.
The function returns a list containing the generated data and relevant information about the simulation.
}
\examples{
\donttest{
# Generate data from the GC spatial regression model with geospatial dependency

data <- rGCgeo(n = 100, alpha = 1, beta0 = 0.3, beta = c(-0.5, 0.5),
sigma = 1, range = 2)

# View the generated data
print(data)
}

}
